Movement, Shooting and Endless Tile Patterns
In this commit i added the movement mechanic to the player character. Allowing the player character to go up and down in positive and negative gravity.

Allowing an arm to rotate so players can shoot with the mouse button

and allowing the game to spawn random prefabed tiles so there are different types of platforms that can be made to change gameplay.