Now there are 4
So now for 3D assets we have a shark, A whale, A StarFish and sea turtle. If i have the time i do play to give them small animations so they look like they're swimming. The shark already has one. Next i'll be making a clown fish and a barracuda. Then i'll be moving on the sea life like coral and sea-weed