The Beginning of our AI
Okay so in this commit i started creating the AI of our game. The AI has the ability to have a cone of sight allowing for different game states that will be implemented soon. These are dormant which is just idle really. Hunting which is trying to find the player and alert where the AI will engage in combat. Hunting there will also be a random gun fire later.
The AI stops when it sees the player in the angle that i have set up with the cone of sight. There is also a concept art i this commit too.

Next time i will be making the AI move more effectivly and changing movement types depending on behaviour states. As well as shooting the player and avoiding obsticles when moving from location to location.