Update 2 AI Script
Now the AI changes behaviour and goes back to normal with the conditions i have set. The AI can also detect walls however cannot avoid them currently as i9 have not set up the movement correctly for it to do so. This is what i will be working on next and also working on the hunting movement