Noise Detection
So because we are still in the early stages of developing this game i have made a very simple noise detection logic to the AI. The AI will go to where the player shot their weapon and investigate what happened. Its simple. The player shoots - The NPC looks in that direction and moves towards it. the AI will go in the area the object will destroy and then the AI goes back to regular movement. There will need to be some revisiting to this logic as we cant have all NPCs in one level gpoing towards a point in the world. Maybe a radius can be implemented but in the future