Enemy FOV
this is just a quick commit to show the FOV script works with the current AI logic. I have found a bug which i have explained in the FOV script. Just a boolean not ticking back but right now  this is not to important

Next im diving into Hunting movement and maybe editing Players gun mechanics and adding headshot logic and aiming down scope