Combat and AI
Okay so in this commit i made a Headshot logic where a cube falls off the AI character if its the last thing you hit before the AI death. Also the AI now has health and the PC can cause damage with the different firing types which now include Burst Firing. Their is also now an explosive bullet.

The AI behaviours for hunting need to be implemented then the main elements are finished are are ready for Debugging and improvments.