Pulse Light Controller and Drawers
Okay so in this commit i made a Light Pulse system which manipulates a light component added on start which we can change almost all the light variables that are on the light component. Also i made a custom editor system which allows us to hide or make variables inactive until we click booleans in the script to reveal or activate them. I did this as there are 4 lighting types which have different variable types. Area light type has Width and Height for the size of the light zone and it also has disc which uses radius. Instead of cluttering the main logic i used the custom editor to hide them so it does not get confusing.