LWRP, Reatime CSG Render transfusion, Shaders
In this commit i installed the lightweight render pipeline which allows us to make shaders in shader graph. This did make problems with CSG realtime as the materials were not in sync with the Pipeline so i changed the matterials to the right settings. I also made 2 shaders that i am happy with. A tron looking shader which moves with time. as well as a outline object shader however this needs improvment as in the scene you can see the shaders effect however during runtime you cannot. I also cleaned up the code allowing for a more effective way to debug. A combination of keys can be pressed in which allows variables to be unhidden and seeing the private variables. this is for testing and to see if everything is working correctly behind the scenes. It also makes it easier to find bugs during runtime