Stop patrolling and AI Fire detection
So in this commit i made a simple stop to the AI patrol movement. Although it is very basic it stops the AI when its on its patrol which can be edited outside the script. It also allows for more gameplay states. like flanking from the back etc.

I also made a simple detection where the player understands where they are being shot from which will be shown in UI. So far we can detect the front and back