Weapon Sway | Central Aim | Script clean up
In this commit the Weapon now sways left right up and down depending on the direction the mouse is moved. Which allows for the object to drag slowly when player rotates. Also making the Aim shoot where the iron sight is. Which was done by shooting at the camera and aligning the aiming position to the fire position which now works for everywhere we aim. I also cleaned some of the scripts up so that Components are found through the scene on runtime and through the asset folder. Also allowing the AI to have a bool which controls if the AI stops and also a way for AI to move even if its stop logic is not finished.