Melee Attack and shoot fixes
I attached the pistol mesh to the PC so when can remove the placeholder gun. I also created a melee attack in which hurts the AI its very very basic tho. I also fixed some wierd bug that was happening in the Gunscript which didnt allow the bullets to go down as an if statement was clashing with the rest of the logic