Code Cleanup
In this commit i made sure the AI does not detect you as soon as you press shift! There was a weird behaviour in which when you held shift but did not move the AI would still know where you were. So there is a cooldown for this now. The AI Also goes straight into alert after being shot twice. As if something in the world like an enemy is being shot it should know after to bullets where the Player is so it has a fighting chance.

I did also clean up the AI Scripts update method as it was getting very untidy and seperated core logics into new functions. There is also an AI stop now in which allows the AI when alert to stop before ramming into the player. I only did this functionality for alert states as when searching we dont want the AI to stop at a distance between itself and the player position and when dormant the AI is patrolling in which if the player is in view then the states go through their cycle to dertermine the tides of gameplay.

I also increased the size of the FOV for the ai when searching dormant alert etc and set up variables in which can be edited to find the best Radius for the AI. I also did this for the speed changes per state so we can speed up the AI when it changes states.