More animations
Okay so the 3D mesh i made can now walk and stop while looking around. The walking and looking around are inside a blend tree in which allows for the searching behaviour to be shown and the stopping during patrol. I did not set this up for the NPC as i dont think its appropriate right now for the proof of concept. I also dont want to ruin the current prefab enemy that works. Also, i should add that talking with the team the mesh i made could be used for the NPC instead of throwing it away after the actual meshs are made.