NPC Talking and Ring effect
Okay so i made a script which will allow the player character to talk to an NPC. it also uses the Type Writer effect we have too. I used and made a particle effect which loops in a ring to show the player the zone they need to be in to talk. I also did this so it cates the players attention