Level Manager
In this commit i deicded to make a level manager which would be handy later in the game. The level manager will be used to balance the game and contain information which can cause restrictions with abilities, also containing how many objects are in one scene so we can play around and turn them off to save CPU process. The Level manager also has a spawner to just check if there is a level mnager already in the scene. if the level manager is not in  the scene then it will spawn it for us