Gun Recoil, Running Sway Effect, minor fixes
So in this commit i made a recoil for the guns that will be in the game. when the player shoots something the gun goes up. Next the Running Sway is an effect in which pans the pistol left when running.

I also fixed minor issues like the dialogue script input bug. which made it so that the typewriter effect we have for the text gets gumbled up if you press the talk key to much in a matter of seconds. This was fixed by adding a input cooldown.

I also fixed boolean logic for the new logic in our the weapon play being the recoil and the running sway. while also making the explosive bullets make objects fly further away.