Better recoil and button manager conditions
So i implemented a better way to show recoil instead of the gun just bouncing up. The recoil now works with rotation while changing the weapons position. There is also a recoil for aiming in which values can be changed in the Unity Inspector.

I made simple if statements in the button manager as some functions and variables were becoming errors. this is because of the correct menus not being in levels so a quick condition fixed the error.