New Slope checker for PC, Bug fixes and debug conditions
In this commit i made a slop logic which allows the player to go up simple slopes (Slopes that are not steep) while also allowing the player to go up steps. This was needed as the players speed kept going down on slops and could get caught on collision of clipping objects. The logic also helped with gravity issues. To much gravity made the player move very slow and then altering the player speed made the PC to fast. Player now goes between 2 gravitys. normal world gravity -9.81 or a gravity we can set in the inspector.

I also fixed the constant reload animation problem with a simple condition. The gun could be reloaded on max ammo which was odd so this can now no longer happen. Also i set some conditions up on start just in case some logic does not apply to the current state of the game.

My next move is to fix all the bug that have been found while also talking withh QA when done to make sure they have been fixed in full and balancing gameplay if time is relevent.