Asset File Restructure / GameManager SetUp
In this commit i restructured the asset folder. I only dragged all the assets into a folder called "Resources" in which allows me to find data from script. so it i want to find a prefab in the assets without finding it on runtime i can do so. I did this as i needed to access the Player prefab and UI prefabs (Pause/Restart and PlayerUI) as i was making a GameManager.

The GameManager is a way to keep all the scenes updated at all times. While trying to repair bugs with my peer i noticed that everyone had outdatted versions of the game. The Game Manager now prevents this as it spawns the correct UI and player character. This will be a great way to stop bugs in my peers scene and allow them to work without struggle.

Now i will be working with the QA lead to fix bugs and balance gameplay