Repsawn Bug repair and AI condition
This commit was just a bug repair in which the QA had found where the player could not respawn. This is now fixed. Just as much as because our player hes death logic we need the AI to change state when the player dies and go back to being dormant which is in effect.