Level Manager Upgrade
Thinking that i had fixed the collision issue, although what i did worked my solution was not good as there will be a lot of walls in the game in which not all colliders will be esculated in their Y size of the component.

I am now instead making a Wall detecter in which allows the level manager to detect the nearest wall in the world that is around the player and what is considered as a wall. The walls are found by having an object layered as Wall.

What will happen is when the player is near to the closest wall in their current area the player will yield their gun in a position where the player cannot climb objects using the guns collision.
 - I also deleted a file that had nothing isnide it