Wall Collision Bug resolved
Okay so there was a bug in which has now been fixed. The bug was the player was able to use the gun to climb objects in the environment. These objects are not meant to be climable. The first fix being a collision increase on the Y was not a suitable enough fix as there will be a lot of walls or objects like this in the game.. So there could be riskof the collider not being high enough.

So instead the Level Manager was made to spawn important objects if they are not already in the scene. While also using the Level Manager as a space to interrupt core functionality.

The gun now goes off screen when near an object which its layer is wall. This is very important to have set up as the level manager also finds these walls through an array and it makes it easier to see which wall is the closest to the player.

The player can now no longer climb!