Slop bug repair, UI pop up text and AI headshot fix
In this commit i was chasing after bugs that the QA had reported while also making some player feedback when shooting enemies.

When you shoot an NPC there will be a pop up damage text which shows the AIs current health. While also fixing the player down slope checker as it was causing some jiggerty motion which was noticed but the QA. I also fixed the AIs headshot logic. there was an error occuring so this now longer shows in the console. The error was saying that the object could not be deleted so the destroy exacution changed.

I also made sure the level manager found all the Text being spawned so it can delete it withing time to keep the game free of objects that have no more use