Hit Detection Feedback, Bug repairs, Player Movement
In this commit there was still an issue with the slop detection, however it was with going down them when running. This is now resolved and works smoothly. I also made sure that the hit indication UI is no longer shown on screen all the time but isntead when the player is shot from a direction there is an image which spawns to notify the player while also fadding away and deleting itself in time.

I fixed an issue with the AI headshot logic which ones again related to the slop checks. Because the Physics.gravity project settings were being manipulated it was applying for all instead of just the player. So the heads would move through the floor collision with to much speed and would not push back with force.

Now the player has a new gravity check which works well and does not have any on going effects later on. I also made sure the player could not shoot when reloading as this was something i found when pressing r to reload and spamming shoot,