Added some script to service gameplay
I created a script that allows to move an object up and back to the origin point, this can be used to move object in a level as well as to enrich the environment. an example is the charging port sign, which now hovers. i made changes to the rotate script for the same reasons.


I've started a new scene to test the start menu.

I am now working on my individual level, as i consider the work on the hub zone to be completed from a design point of view