Healthbar exposure and repair, bloodlock script
I made a blood lock script which allows us to manipulate objects in the scene depending if an array of other objects have died.

We can Activate/DeActivate a single object or an array of them. I also fixed the healthbars transparent behaviour as well as exposing an array of colors which we can edit for each stage of the players healthbar feedback