Simple Font Changer
In this commit a i made a simple Font changer which will be improved later. it allows players to pick a font from the start menu and else where in the game. The drop down value allows the fonts to cycle in an array. This also relates to TextMeshPro. However, i think this could get difficult to have a consistent font later on.