Gun Upgrades & Weapon sway update & Footsteps & Font change
So in this commit i made it so that the different logic of the gun can not be used until unlocked. In which i am talking about the explosive bullets, Full auto and burst fire shooting. These will not activate until the player finds the pickups

I also updated the weapon sway so now the gun goes up and down when running.

I also added footsteps for when the player is walking and running while also commenting the UI text logic inside the font changer in the start menu so it only focuses on TextMeshPro

Some Text components in the UI will need to be changed to support the text mesh pro font changer