Background, Lighting, Jump
the background now has original 3d models and planes to represent buildings far away. the start position now has a barrier to guide the player towards the first exposition delivery. the exposition cube now has a 3d model of an arcade cabinet as housing. the global lighting has been improved, now npcs are visible in the shadows. a jump script has been added to the PC, now the player can jump, however, it might have broken the AI. the pc running sway  now bring the gun towards the middle of the screen and rotates a bit, implying that the gun is not available during that action