Reload Audio Sync, AI knockback and Health-Ammo pickups
In this commit i made it so that the Ai can be knocked back when shot in the head or body. While also allowing for a health or ammo pickup to be used in the game which have been made into prefabs too. The reladion animation for the players gun now has better synced audio with the reloading sound effect