AI Repair
After a lot of time trying to repair the AI rotation and movement. I found that the Prefab alignments were wrong. The NavMesh was flipping the Prefab so it would face to the correct position. However, because of the confusition of the mesh being a child of the prefab everything looked fine. The prefab now works and can be refined with its movement and shooting.