Knockback percentage control and PC accuracy
The knockback no longer happens with each shot that the player does to the AI. The knockback is generated when a random value is met. This was it leaves room for critical hit dealing more damage. However, the knockback will need to be balanced and edited for each enemy type.  The players now cannot be dead accurate with the shooting only with the single shot firing type. Damage will need to be revisited for each shooting mode and this method can be restructured if needed.

Next i will be editing the level manager and also making sure the game builds correctly