Build fixes
On build the player character would not move fast and the camera rotation was not acceptable as it was to slow. This is now longer an issue. The pulse light i discovered that some aspects of the LightType settings in unity do not like being manipulated on build so this will need to be set up in the engine.