Conversation Script repair
The issue with the script was the Type Writer effect was being called to fast and the story events were being called at the wrong time. Allowing the story event to be called on trigger allows it to be runned one time only.

The script we need to be revisted so Text resets all the time as a constant check.

Co-Authored-By: thatdevboi <thatdevboi@users.noreply.github.com>