Font Manager Repair
Font Manager was having a hard time finding variables while trnasitioning through scenes. This is no longer an issue.

The Font manager could not recover the data it needed like the drop down menu and the options canvas. It was also finding errors in which restricted lines of source code being exacuted.