Array Element string name drawer, Audio editor w multiple edited mixers
I made mixers which allow us to priorities volume of certain sounds more than others. They are their master sound in which controls an Audio Source volume. I made sure we can alter these Audio mixers with a slider in the start menu. In which will allow players more accessibility when altering settings. This will be the same when controlling the sensertivity of the mouse rotation, this will be made later when we decided on the smooth spot values. I also made a handy Drawer which edits the inspector. This allows us to name Array Elements so we are certain what should be in each array. We can name them in the source code by just writing strings. I have showed examples of this in the UI_Manager. I also changed the DropDown menu for the font so it supports TextMeshPro.