Options Menu Revamp
I revamped the options menu so there are sub menus that covers the audio/controls and visuals. I also changed the idea of using multiple sound mixers as the old options menu layout was unappealing to the team and looked confusing. The new mixer carries groups in which we can manipulate different groups audio from the new Options menu layout. We can also keep adding to the menu. The buttons of the sub menus change color so its easy to see what menu is active too. Next i will be revisiting the movement and syncing the new animations to the movement