Player Movement instant stop and Syncing animations and sounds
I synced the sounds with the animations for reloading and moving. I used animation events that just use a Function to be called and does as required. I also made sure when the player releases the movement keys w, a, s, d then the player loses all velocity and stops instantly.