Movement Revamp
For the movement revamp i had to clamp the players Velcoity as we are changing this depending on the current type of movement Walking/Running. I then also had to put the movement function into FixedUpdate so it did not have any delays. From what i can see the player stops when the key is released