Level Design
I added small details to my level. The idea at hand is to make a very religious location a place of slaughter. As the robots is manipulated hurting people at a church can be an impacting event. I tried to make the church the most noticable item in the level while also thinking about how are people getting here. I settled on a simple train platform. The level will however need a lot of different meshs to feel closer to reality. Meshs like tents, Trains, Smaller buildings like shops. Areas to be used as covers. Also allowing for interior design of the church so players can enter the church. I made a prefab of my level so it can be added to the main scene