Audio Submenu, Fist, In-game UI
The audio sub-menu now is centred and had a darker background, allowing the use of white fonts.

the fist model made by Nad  are now in the pc prefab

the in game ui now is cleaner to read. several objects has been moved to a more user friendly position. may need testing