Level design
Worked on my level idea and came up with a big buidling to aim for at the end. Added an area for the large building and increased the size of the outside ones as suggested last session.
Enemies could not be added in this prefab as having them attacthed to a parent breaks their AI.