Player Jumping, Code Clean
In this commit i cleaned the inspector window of variables that did not need to be seen, also removing variables that was no longer used for the games direction.

I however, made a jumping method. The method we had in the game would fight with our built in gravity system in which the player would barly jump at all. This new jumping system supports the gravity method so players can jump to a certain Maximum distance and then decend with gravity. The jump can be manipulated through the variables i have set up in the class.  It will come under the Header Jumping in the Player Controller class attached to the PC prefab.