Bug fixes
I started to fix some of the bugs that are on the trello. I allowed for a boolean to stop players from shooting when in and exiting the pause menu. The bool stops the input and this bool can be used for other functions in the future. I also made sure the mini boss range was not to large, as well as i made sure the mini boss could not just shoot in a dormant state. It now only shoots in the Chasing and fazing states. I also made sure the sound was synced with the walking again as the clip for walking was changed.