Zipline and dialogue alternative
I made a zipline mechanic which can be used in the levels to change verticality. I also made sure we can decide what objects show dialogue and what objects are for voice acting. I also fixed an issue where the voice acting would not loop from 3D sound to 2D sound. If the player was in 2D then the audio would never repeat. so i made sure to add a loop