Bug repairds
i fixed and balanced the players jump as there was a lot of weird behaviour. For example the maximum height the player could jump needed to be changed as players could reach the height limit and would not jump. Or players could jump extremly high when spawning in a high area which then players would need to go down. I also allowed the Zipeline to be used in a loop. The reset i had put into place was not functioning correctly. This is now no longer a problem.