Bug repair
I fixed a few minor bigs in the game and also will be trying to fix a bug where a bullet trail effect follows where the player faces. This will be my next thing i want to do as well as working on balancing stats like damage, speed, rotation etc. I want to balance this within how it shows in the engine and how it shows within the build